% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcc_bioclim.R
\name{cmcc_bioclim}
\alias{cmcc_bioclim}
\title{A function to parse historic BIOs from CMCC-BioClimInd.}
\usage{
cmcc_bioclim(bry = NULL, path = NULL, nm_mark = "clip", return_stack = TRUE)
}
\arguments{
\item{bry}{(sf or sp) The boundary to mask the data.
if NULL, no clip to the global map. The default is NULL.}

\item{path}{(character) The path to save the downloaded imagery.
If NULL, then use the current working directory. The default is NULL.}

\item{nm_mark}{(character) the name mark of clipped images.
The default is "clip". It would be ignored if bry is NULL.}

\item{return_stack}{(logical) if TRUE, stack the imagery together and return.
If the area is large and resolution is high, it is better not to stack them.
The default is TRUE.
#' @return stars if return_stack is TRUE.
The images would be saved as a single file.}
}
\description{
This function allows you to parse historic CMCC-BioClimInd
bioclimatic variables with a setting of boundary and a few other options.
}
\examples{
cmcc_bioclim()
cmcc_bioclim(return_stack = FALSE)
}
\references{
Noce, Sergio, Luca Caporaso, and Monia Santini.
"A new global dataset of bioclimatic indicators."
Scientific data 7.1 (2020): 1-12.
\url{https://doi.pangaea.de/10.1594/PANGAEA.904278?format=html#download}
}
