% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_pa.R
\name{convert_to_pa}
\alias{convert_to_pa}
\title{A function to convert suitability to presence-absence map.}
\usage{
convert_to_pa(
  suitability,
  method = "logistic",
  beta = 0.5,
  alpha = -0.05,
  a = 1,
  b = 0,
  species_prevalence = NA,
  seed = 1,
  visualize = TRUE
)
}
\arguments{
\item{suitability}{(stars or RasterLayer) The suitability raster.}

\item{method}{(character) The conversion method, must be one of \code{'threshold'},
logistic, and linear.}

\item{beta}{(numeric) Works for threshold or logistic method. If method is
threshold, then beta is the threshold value to cutoff. If method is logistic,
it is the}

\item{alpha}{(numeric) Works for logistic method.}

\item{a}{(numeric) Works for linear method. It is the slope of the line.}

\item{b}{(numeric) Works for linear method. It is the intercept of the line.}

\item{species_prevalence}{(numeric) Works for all three methods. It is the
species prevalence to classify suitability map.}

\item{seed}{(integer) The seed for random progress.}

\item{visualize}{(logical) if TRUE, plot map of suitability, probability of
occurrence, and presence-absence. The default is TRUE.}
}
\value{
(PAConversion) a list of suitability map, probability of occurrence,
conversion details, and presence-absence map.
}
\description{
Use threshold-based, logistic or linear conversion method to
convert predicted suitability map to presence-absence map.
}
\details{
Multiple methods and arguments could be used as a combination to do
the conversion.
}
\examples{
convert_to_pa(suitability)

}
\references{
https://github.com/Farewe/virtualspecies/blob/master/R/convertToPA.R
}
