% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_reduce.R
\name{dim_reduce}
\alias{dim_reduce}
\title{A function to remove auto-correlation between raster layers.}
\usage{
dim_reduce(
  img_stack = NULL,
  threshold = 0.5,
  preferred_vars = NULL,
  samples = NULL
)
}
\arguments{
\item{img_stack}{(Stars or RasterStack) The image stack to work on.}

\item{threshold}{(numeric) The number of the threshold to reduce.
The default is 0.5.}

\item{preferred_vars}{(vector of character) The preferred variables in order
in dimension reduction. The preferred variables will move to the beginning
before the reduction. So make sure they are placed in order. Furthermore,
setting preferred_vars does not guarantee they can survive, for example, one
preferred variable that is placed later has strong correlation with former
preferred variable.}

\item{samples}{(sf or sp) The samples to reduce dimension.
If NULL, the whole rasterstack would be used. The default is NULL.}
}
\value{
ReducedImageStack.
}
\description{
This function allows you to reduce dimensions of raster layers
based on pearson correlation and a user-defined threshold.
}
\examples{
worldclim <- worldclim2(var = "bio")
dim_reduce(worldclim)
}
