% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_po.R
\name{evaluate_po}
\alias{evaluate_po}
\title{A function to evaluate the model based on presence-only data.}
\usage{
evaluate_po(model, occ_pred, bg_pred = NULL, var_pred, visualize = FALSE)
}
\arguments{
\item{model}{(isolation_forest) The extended isolation forest SDM.}

\item{occ_pred}{(vector) the vector contains predicted values at occurrence locations.}

\item{bg_pred}{(vector) the vector contains predicted values with same number
of background points.}

\item{var_pred}{(vector) the vector contains predicted values of the whole area.}

\item{visualize}{(logical) If TRUE, print message and plot the curves.
The default is FALSE.}
}
\value{
(POEvaluation) a list of evaluations
}
\description{
This function will calculate Contrast Validation Index (CVI),
continuous Boyce index (CBI) and AUC_ratio.
CVI is the proportion of presence points falling in cells having
a threshold (0.5 here) habitat suitability index minus the proportion of
cells within this range of threshold of the model.
Here we used varied thresholds: 0.25, 0.5, and 0.75.
AUC_background uses background values as pseudo-absence values
ROC curve plots the proportion of presences falling above a
range of thresholds against the proportion of cells falling
above the range of thresholds. The area under the modified
ROC curve was then called AUC_ratio.
}
\examples{
evaluate_po(mod, occ_pred = occ_pred, var_pred = var_pred)
}
