% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_cmcc_bioclim.R
\name{future_cmcc_bioclim}
\alias{future_cmcc_bioclim}
\title{A function to parse future BIOs from CMCC-BioClimInd.}
\usage{
future_cmcc_bioclim(
  bry = NULL,
  path = NULL,
  esm = "CMCC-CESM",
  rcp = 85,
  interval = "2040-2079",
  nm_mark = "clip",
  return_stack = TRUE
)
}
\arguments{
\item{bry}{(sf or sp) The boundary to mask the data.
if NULL, no clip to the global map. The default is NULL.}

\item{path}{(character) The path to save the downloaded imagery.
If NULL, then use the current working directory. The default is NULL.}

\item{esm}{(character) The option for  Earth System Models (ESMs).
Should be one of "CMCC-CESM", 'GFDL-ESM2M', 'HadGEM2-ES',
'IPSL-CM5A-LR', 'MIROC-ESM-CHEM', 'NorESM1-M'.
The default is CMCC-CESM.}

\item{rcp}{(numeric) The option of Representative Concentration Pathways (RCPs).
Should be 45 or 85. Only 85 is available for CMCC-CESM.
The default is 85.}

\item{interval}{(character) The option for time interval.
Should be one of "2040-2079", "2060-2099".
The default is 2040-2079.}

\item{nm_mark}{(character) the name mark of clipped images.
The default is "clip". It would be ignored if bry is NULL.}

\item{return_stack}{(logical) if TRUE, stack the imagery together and return.
If the area is large and resolution is high, it is better not to stack them.
The default is TRUE.
#' @return stars if return_stack is TRUE.
The images would be saved as a single file.}
}
\description{
This function allows you to parse future CMCC-BioClimInd
bioclimatic variables with a setting of boundary and a few other options.
}
\examples{
future_cmcc_bioclim()
future_cmcc_bioclim(return_stack = FALSE)
}
\references{
Noce, Sergio, Luca Caporaso, and Monia Santini.
"A new global dataset of bioclimatic indicators."
Scientific data 7.1 (2020): 1-12.
\url{https://doi.pangaea.de/10.1594/PANGAEA.904278?format=html#download}
}
