% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_worldclim2.R
\name{future_worldclim2}
\alias{future_worldclim2}
\title{A function to parse the future climate from worldclim version 2.1.}
\usage{
future_worldclim2(
  var = "tmin",
  res = 10,
  gcm = "BCC-CSM2-MR",
  ssp = "ssp585",
  interval = "2021-2040",
  bry = NULL,
  path = NULL,
  nm_mark = "fut",
  return_stack = TRUE
)
}
\arguments{
\item{var}{(character) The option for the variable to download.
Should be one of tmin, tmax, prec, bioc.
The default is tmin.}

\item{res}{(numeric) The option for the resolution of image to download.
Should be one of 2.5, 5, 10.
The default is 10.}

\item{gcm}{(character) The option for global climate models.
Should be one of "BCC-CSM2-MR", "CNRM-CM6-1","CNRM-ESM2-1", "CanESM5",
"GFDL-ESM4", "IPSL-CM6A-LR","MIROC-ES2L", "MIROC6", "MRI-ESM2-0".
The default is BCC-CSM2-MR.}

\item{ssp}{(character) The option for Shared Socio-economic Pathways.
Should be one of "ssp126", "ssp245", "ssp370", "ssp585".
The default is ssp585.}

\item{interval}{(character) The option for time interval.
Should be one of "2021-2040", "2041-2060", "2061-2080", "2081-2100".
The default is 2021-2040.}

\item{bry}{(sf or sp) The boundary to download the data.
If NULL, no clip to the downloaded files.
The default is NULL.}

\item{path}{(character) The path to save the downloaded imagery.
If NULL, the use the current working directory.
The default is NULL.}

\item{nm_mark}{(character) The prefix of the file names to download.
The default is fut.}

\item{return_stack}{(logical) If TRUE, stack the imagery together and return.
The default is TRUE.
#' @return stars if return_stack is TRUE.
The images would be saved as a single file.}
}
\description{
This function allows you to parse worldclim version 2.1 future climatic
files with a setting of boundary and a few other options.
}
\examples{
future_worldclim2("tmin", 10, "BCC-CSM2-MR",
"ssp585", "2021-2040", return_stack = FALSE)
}
\references{
\url{https://worldclim.org/data/index.html}
}
