% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_response.R
\name{independent_response}
\alias{independent_response}
\title{A function to calculate independent responses of each variables.}
\usage{
independent_response(model, var_occ, variables, si = 1000, visualize = FALSE)
}
\arguments{
\item{var_occ}{(data.frame, tibble) the data.frame style table that
include values of environmental variables at occurrence locations.}

\item{variables}{(RasterStack, RasterLayer or stars)
the stack of environmental variables. Could be raster or stars object.}

\item{si}{(integer) the number of samples to generate response curves.
If it is too small, the response curves might be biased.
The default value is 1000.}

\item{visualize}{(logical) if TRUE, plot the response curves.
The default is FALSE.}
}
\value{
a list of data.frame of response and variable values.
The response values correspond to suitability of this single variable.
}
\description{
This function allows you to calculate the independent responses
of each variables within the model.
}
\details{
The curves show how each environmental variable independently
affects the modeling prediction. The curves show how the predicted result
only using this variable changes as it is varied.
}
\examples{
independent_response(model = mod$model, variables = env_vars)
}
