% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotree_sdm.R
\name{isotree_po}
\alias{isotree_po}
\title{Function to run extended isolation forest as SDM.}
\usage{
isotree_po(
  occ,
  occ_test = NULL,
  occ_crs = 4326,
  variables,
  ntrees = 100L,
  sample_size = NA,
  sample_rate = NA,
  ndim = 0L,
  ...,
  response = TRUE,
  check_variable = TRUE,
  visualize = FALSE
)
}
\arguments{
\item{occ}{(data.frame, sf, SpatialPointsDataFrame)
The occurrence dataset for training.
There must be column x and y for coordinates if it is a data.frame.}

\item{occ_test}{(data.frame, sf, SpatialPointsDataFrame)
The occurrence dataset for independent test.
There must be column x and y for coordinates if not NULL and it is a data.frame.
If NULL, no independent test will be used. The default is NULL.}

\item{occ_crs}{(numeric or crs) The EPSG number or crs object of occurrence CRS.
The default value is 4326, which is the geographic coordinate system.}

\item{variables}{(RasterStack or stars) The stack of environmental variables.}

\item{ntrees}{(integer) The number of trees for the isolation forest.
The default is 100.}

\item{sample_size}{(integer) A number for sampling size in \verb{[0, nrow(occ)]}.
The default is NULL.}

\item{sample_rate}{(numeric) A rate for sampling size in \link{0, 1}.
The default is NULL. Only set either \code{sample_size} or \code{sample_rate}.}

\item{ndim}{(integer) ExtensionLevel for isolation forest,
it must be no smaller than the dimension of environmental variables.
When it is 0, the model is a traditional isolation forest, otherwise the model
is an extended isolation forest. The default is 0.}

\item{...}{Other arguments that \code{isotree::isolation.forest} needs.}

\item{response}{(logical) if TRUE, generate response curves.
The default is TRUE.}

\item{visualize}{(logical) if TRUE, generate the essential figures related to
the model. The default is FALSE.}

\item{check_variables}{(logical) if TRUE, check the variable importance.
The default is TRUE.}
}
\value{
(isotree_presence_only) a list of model, occ var_train, pred_train,
pred_test,  prediction, responses, eval_train, and eval_test.
}
\description{
Call isolation forest and its variations to do species distribution modeling.
}
\examples{
isotree_po(occ = occ, variables = env_vars, ntrees = 500,
sample_size = 400, ndim = 0)

}
