% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mainland_africa}
\alias{mainland_africa}
\title{Boundary of mainland Africa}
\format{
A \code{\link[sf:sf]{sf}} with one rows and 2 fields
\describe{
\item{name}{(\code{character}) The name of the polygon: Africa}
\item{area}{(\code{\link{units}}) The united number of the overall area in
km2. This is not a consensus area, but just a calculated area under this
resolution.}
\item{geometry}{(\code{\link[sf:sfc]{sfc}}) The simple polygon feature of the
boundary}}
}
\source{
\code{rnaturalearth}
}
\usage{
mainland_africa
}
\description{
The overall continental boundary of mainland Africa queried from
\code{rnaturalearth} and get processed.
}
\keyword{datasets}
