% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_response.R
\name{marginal_response}
\alias{marginal_response}
\title{A function to calculate marginal responses of each variables.}
\usage{
marginal_response(model, var_occ, variables, si = 1000, visualize = FALSE)
}
\arguments{
\item{model}{Any predictive model. In this package, it could be isolation_forest.
In this package, it is isolation_forest.}

\item{var_occ}{(data.frame, tibble) the data.frame style table that
include values of environmental variables at occurrence locations.}

\item{variables}{(RasterStack, RasterLayer or stars)
the stack of environmental variables. Could be raster or stars object.}

\item{si}{(integer) the number of samples to generate response curves.
If it is too small, the response curves might be biased.
The default value is 1000.}

\item{visualize}{(logical) if TRUE, plot the response curves.
The default is FALSE.}
}
\value{
(MarginalResponse) a list of data.frame of response and variable values.
The response values correspond to suitability of this single variable.
}
\description{
This function allows you to calculate the marginal responses of
each variables within the model.
}
\details{
The curves show how each environmental variable affects the modeling
prediction. The curves show how the predicted result changes as each environmental
variable is varied while keeping all other environmental variables at average
sample value. The curves might be hard to interpret if there are strongly correlated
variables. The users could use \code{dim_reduce} function to remove the strong correlation
from original environmental variable stack.
}
\examples{
marginal_response(model = mod$model, variables = env_vars)
}
