% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detect.R
\name{outlier_detect}
\alias{outlier_detect}
\title{A function to detect outliers from occurrence.}
\usage{
outlier_detect(env_vars, remove_outlier = FALSE)
}
\arguments{
\item{remove_outlier}{(logical) the option to remove detected outliers or not.
The default is FALSE.}

\item{occ}{(data.frame) The occurrence data frame for training.
There must be column x and y for coordinates. This argument is required}
}
\description{
This function allows you to detect outliers from occurrence dataset.
}
\examples{
outlier_detect()

}
