% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.EnvironmentalOutlier}
\alias{plot.EnvironmentalOutlier}
\title{Function to plot suspicious outliers in an observation dataset.}
\usage{
\method{plot}{EnvironmentalOutlier}(x, overlay_raster = NULL, pts_alpha = 0.5, ...)
}
\arguments{
\item{x}{(EnvironmentalOutlier) The PAConversion object to plot.
It could be the return of function \code{suspicious_env_outliers}.}

\item{overlay_raster}{(RasterLayer or stars) The environmental raster to plot
together with points.}

\item{pts_alpha}{(numeric) the alpha used by ggplot2 to show points.}

\item{...}{Not used.}
}
\description{
Display observations and outliers in a dataset.
}
\examples{
plot(suspicious_outliers)

}
