% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.evaluation_po}
\alias{plot.evaluation_po}
\title{Function to plot presence-only evaluation.}
\usage{
\method{plot}{evaluation_po}(x, ...)
}
\arguments{
\item{x}{(evaluation_po) The presence-only evaluation object to plot.
It could be the return of function \code{evaluate_po}.}

\item{...}{Not used.}
}
\value{
a patchwork of ggplot figure of AUC_ratio, AUC_background and CBI.
}
\description{
Display informative and detailed figures of continuous Boyce
index and AUC curves.
}
\examples{
plot(evaluation_po)

}
