% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.IndependentResponse}
\alias{plot.IndependentResponse}
\title{Function to plot independent response curves.}
\usage{
\method{plot}{IndependentResponse}(x, target_var = NA, smooth_span = 0.3, ...)
}
\arguments{
\item{x}{(IndependentResponse) The independent response curve object to plot.
It could be the return of function \code{independent_response}.}

\item{target_var}{(vector of character) The target variable to plot. It could be
NA. If it is NA, all variables will be plotted.}

\item{smooth_span}{(numeric) The span value for smooth fit in ggplot2.
When it is 0, no smooth applied. The default is 0.3.}

\item{...}{Not used.}
}
\value{
ggplot2 figure of response curves
}
\description{
Plot independent response curves using ggplot2.
}
\examples{
plot(independent_responses)

}
