% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.marginal_response}
\alias{plot.marginal_response}
\title{Function to plot marginal response curves.}
\usage{
\method{plot}{marginal_response}(x, smooth_span = 0.3, ...)
}
\arguments{
\item{x}{(marginal_response) The marginal response curve object to plot.
It could be the return of function \code{marginal_response}.}

\item{smooth_span}{(numeric) The span value for smooth fit in ggplot2.
When it is 0, no smooth applied. The default is 0.3.}

\item{...}{Not used.}
}
\value{
ggplot2 figure of response curves
}
\description{
Plot marginal response curves using ggplot2.
}
\examples{
plot(marginal_responses)

}
