% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.PAConversion}
\alias{plot.PAConversion}
\title{Function to plot results of conversion to PA.}
\usage{
\method{plot}{PAConversion}(x, ...)
}
\arguments{
\item{x}{(PAConversion) The PAConversion object to plot.
It could be the return of function \code{convert_to_pa}.}

\item{...}{Not used.}
}
\value{
a patchwork of ggplot figure of suitability, probability of occurrence,
presence-absence binary map.
}
\description{
Display raster of suitability, probability of occurrence,
presence-absence binary map from PA conversion.
}
\examples{
plot(pa_covert)

}
