% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.VariableAnalysis}
\alias{plot.VariableAnalysis}
\title{Function to plot variable importance.}
\usage{
\method{plot}{VariableAnalysis}(x, ...)
}
\arguments{
\item{x}{(VariableAnalysis) The variable importance object to plot.
It could be the return of function \code{variable_analysis}.}

\item{...}{Not used.}
}
\value{
a patchwork of ggplot figure of variable importance
according to multiple metrics.
}
\description{
Display informative and detailed figures of variable importance.
}
\examples{
plot(variable_analysis)

}
