% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.VariableContribution}
\alias{plot.VariableContribution}
\title{Function to plot variable contribution for target observations.}
\usage{
\method{plot}{VariableContribution}(x, plot_each_obs = FALSE, num_features = 5, ...)
}
\arguments{
\item{x}{(VariableContribution) The VariableContribution object to plot.
It could be the return of function \code{variable_contrib}.}

\item{plot_each_obs}{(logical) The option of plot type. If \code{TRUE}, it will
plot variable contribution for every observation. Otherwise, it will plot
variable contribution violin plot for all observations.}

\item{num_features}{(integer) A number of most important features to plot.
Just work if plot_each_obs is \code{TRUE}.}

\item{...}{Not used.}
}
\value{
ggplot2 figure of Variable Contribution.
}
\description{
Plot variable contribution for target observation separately
or together using ggplot2.
}
\examples{
plot(independent_responses)

}
