% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.VariableDependence}
\alias{plot.VariableDependence}
\title{Function to plot variable dependence obtained from SHAP test.}
\usage{
\method{plot}{VariableDependence}(x, target_var = NA, related_var = NA, smooth_span = 0.3, ...)
}
\arguments{
\item{x}{(VariableDependence) The variable dependence object to plot.
It could be the return of function \code{variable_dependence}.}

\item{target_var}{(vector of character) The target variable to plot. It could be
NA. If it is NA, all variables will be plotted.}

\item{related_var}{(character) The dependent variable to plot together with
target variables. It could be NA. If it is NA, no related variable will be
plotted.}

\item{smooth_span}{(numeric) The span value for smooth fit in ggplot2.
When it is 0, no smooth applied. The default is 0.3.}

\item{...}{Not used.}
}
\value{
ggplot2 figure of dependent curves
}
\description{
Plot variable dependence curves using ggplot2.
}
\examples{
plot(var_dependence)

}
