% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_responses.R
\name{plot_responses}
\alias{plot_responses}
\title{A function to plot the response curves.}
\usage{
plot_responses(response_list)
}
\arguments{
\item{response_list}{(marginal_response or independent_response)
The environmental_response object obtained from modeling.}
}
\value{
ggplot2 figure of response curves
}
\description{
This function allows you to plot response curves using ggplot2.
}
\examples{
plot_responses(response_list = responses)

}
