% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.EnvironmentalOutlier}
\alias{print.EnvironmentalOutlier}
\title{Print summary information from PAConversion object.}
\usage{
\method{print}{EnvironmentalOutlier}(x, ...)
}
\arguments{
\item{x}{(EnvironmentalOutlier) A EnvironmentalOutlier object to be messaged.
It could be the return of function \code{suspicious_env_outliers}.}

\item{...}{Not used.}
}
\value{
The same object that was passed as input.
}
\description{
Display the most general and informative characteristics of
a PAConversion object.
}
\examples{
print(suspicious_outliers)

}
