% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.FormatOccurrence}
\alias{print.FormatOccurrence}
\title{Print summary information from FormatOccurrence object.}
\usage{
\method{print}{FormatOccurrence}(x, ...)
}
\arguments{
\item{x}{(\code{FormatOccurrence}) A \code{FormatOccurrence} object to be messaged.
It could be the return of function \code{\link{format_observation}}.}

\item{...}{Not used.}
}
\value{
The same object that was passed as input.
}
\description{
Display the most general and informative characteristics of
a FormatOccurrence object.
}
\examples{
\donttest{
library(dplyr)
library(itsdm)
data("occ_virtual_species")
obs_formatted <- format_observation(
  obs_df = occ_virtual_species,
  x_col = "x", y_col = "y", obs_col = "observation",
  obs_type = "presence-absence")

print(obs_formatted)
}

}
\seealso{
\code{\link{format_observation}}
}
