% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.PAConversion}
\alias{print.PAConversion}
\title{Print summary information from PAConversion object.}
\usage{
\method{print}{PAConversion}(x, ...)
}
\arguments{
\item{x}{(PAConversion) A PAConversion object to be messaged.
It could be the return of function \code{convert_to_pa}.}

\item{...}{Not used.}
}
\value{
The same object that was passed as input.
}
\description{
Display the most general and informative characteristics of
a PAConversion object.
}
\examples{
print(pa_covert)

}
