% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.POEvaluation}
\alias{print.POEvaluation}
\title{Print summary information from presence-only evaluation object.}
\usage{
\method{print}{POEvaluation}(x, ...)
}
\arguments{
\item{x}{(POEvaluation) A presence-only evaluation object to be messaged.
It could be the return of function \code{evaluate_po}.}

\item{...}{Not used.}
}
\value{
The same object that was passed as input.
}
\description{
Display the most general and informative characteristics of
a presence-only evaluation object.
}
\examples{
print(evaluation_po)

}
