% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ReducedImageStack}
\alias{print.ReducedImageStack}
\title{Print summary information from ReducedImageStack object.}
\usage{
\method{print}{ReducedImageStack}(x, ...)
}
\arguments{
\item{x}{(\code{ReducedImageStack}) A \code{ReducedImageStack} object to be messaged.
It could be the return of function \code{\link{dim_reduce}}.}

\item{...}{Not used.}
}
\value{
The same object that was passed as input.
}
\description{
Display the most general and informative characteristics of
a ReducedImageStack object.
}
\examples{
\donttest{
library(itsdm)
worldclim <- worldclim2(var = "bio")

img_reduced <- dim_reduce(worldclim, threshold = 0.7,
  preferred_vars = c('bio1', 'bio12'))

print(img_reduced)
}

}
\seealso{
\code{\link{dim_reduce}}
}
