% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.variable_analysis}
\alias{print.variable_analysis}
\title{Print summary information from variable importance object.}
\usage{
\method{print}{variable_analysis}(x, ...)
}
\arguments{
\item{x}{(variable_analysis) A variable importance object to be messaged.
It could be the return of function \code{variable_analysis}.}

\item{...}{Not used.}
}
\value{
The same object that was passed as input.
}
\description{
Display the most general and informative characteristics of
a variable importance object.
}
\examples{
print(variable_analysis)

}
