% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.variable_importance}
\alias{print.variable_importance}
\title{Print summary information from variable importance object.}
\usage{
\method{print}{variable_importance}(x, ...)
}
\arguments{
\item{x}{An object to be messaged.}

\item{...}{Not used.}
}
\value{
The same object that was passed as input.
}
\description{
Display the most general and informative characteristics of
a variable importance object.
}
\examples{
print(variable_importance)
}
