% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspicious_env_outliers.R
\name{suspicious_env_outliers}
\alias{suspicious_env_outliers}
\title{Function to detect suspicious outliers based on environmental variables.}
\usage{
suspicious_env_outliers(
  occ,
  occ_crs = 4326,
  variables,
  rm_outliers = FALSE,
  seed = 10,
  ...
)
}
\arguments{
\item{occ}{(data.frame, sf, SpatialPointsDataFrame)
The observation dataset to analyze.
There must be column x and y for coordinates if it is a data.frame.}

\item{occ_crs}{(numeric or crs) The EPSG number or crs object of occurrence CRS.
The default value is 4326, which is the geographic coordinate system.}

\item{variables}{(RasterStack or stars) The stack of environmental variables.}

\item{rm_outliers}{(logical) The option to remove the suspicious outliers or not.
The default is FALSE.}

\item{...}{Other arguments passed to function \code{outlier.tree}.}
}
\value{
(EnvironmentalOutlier) a list that contains \code{sf} points of observations
, outliers and outlier details returned from \code{outlier.tree}.
If \code{rm_outliers} is \code{TRUE}, outliers are deleted from points of
observations. If \code{FALSE}, the full observations are returned.
}
\description{
Call outlier.tree to detect suspicious outliers in observations.
}
\examples{
suspicious_env_outliers(occ = occ, variables = env_vars)

}
