% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_contrib.R
\name{variable_contrib}
\alias{variable_contrib}
\title{A function to evaluate relative importance of each variable.}
\usage{
variable_contrib(
  model,
  var_occ,
  var_occ_analysis,
  shap_nsim = 100,
  visualize = FALSE,
  seed = 10
)
}
\arguments{
\item{model}{(isolation_forest) The extended isolation forest SDM.}

\item{var_occ}{(data.frame, tibble) the data.frame style table that
include values of environmental variables at occurrence locations.}

\item{var_occ_analysis}{(data.frame, tibble) The data.frame style table that
include values of environmental variables at occurrence locations of analysis.}

\item{shap_nsim}{(integer) The number of Monte Carlo repetitions in SHAP
method to use for estimating each Shapley value.}

\item{visualize}{(logical) if TRUE, plot the response curves.
The default is FALSE.}

\item{seed}{(integer) The seed for any random progress. The default is 10.}
}
\value{
(VariableContribution) a tibble of Shapley values of each variable for
each observation.
}
\description{
This function allows you to evaluate relative importance of
each variable within the model using methods: with this variable only,
with all variables except this variables, and permutation feature importance
method. Because the model is presence-only, we only use pearson correlation
with the result of full model as the metrics to rank the variables.
}
\examples{
variable_contrib(model = isotree_po,
var_occ = var_occ, variables = env_vars)

}
