% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_dependence.R
\name{variable_dependence}
\alias{variable_dependence}
\title{A function to calculate variable dependence based on SHAP method.}
\usage{
variable_dependence(
  model,
  var_occ,
  shap_nsim = 100,
  visualize = FALSE,
  seed = 10
)
}
\arguments{
\item{model}{Any predictive model. In this package, it could be isolation_forest.}

\item{var_occ}{(data.frame, tibble) the data.frame style table that
include values of environmental variables at occurrence locations.}

\item{shap_nsim}{(integer) The number of Monte Carlo repetitions in SHAP
method to use for estimating each Shapley value.}

\item{visualize}{(logical) if TRUE, plot the response curves.
The default is FALSE.}

\item{seed}{(integer) The seed for random progress.}
}
\value{
(VariableDependence) a list of data.frame of dependence and variable values.
The dependence values correspond to Shapley value.
}
\description{
This function allows you to calculate the variable dependence
based on Shapley values.
}
\details{
The curves show how each environmental variable independently
affects the modeling prediction. The curves show how the predicted result
only using this variable changes as it is varied.
}
\examples{
variable_dependence(model = mod$model, var_occ = var_occ)
}
