% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldclim2.R
\name{worldclim2}
\alias{worldclim2}
\title{A function to parse worldclim version 2.1 variables.}
\usage{
worldclim2(
  var = "tmin",
  res = 10,
  bry = NULL,
  path = NULL,
  nm_mark = "clip",
  return_stack = TRUE
)
}
\arguments{
\item{var}{(character) The option for the variable to download,
should be one of tvag, tmin, tmax, prec, srad, wind, vapr and bio.
The default is 'tmin'.}

\item{res}{(numeric) The option for the resolution of image to download.
Should be one of 0.5, 2.5, 5, 10 in minute degree.
The default is 10.}

\item{bry}{(sf or sp) The boundary to mask the data.
if NULL, no clip to the global map. The default is NULL.}

\item{path}{(character) The path to save the downloaded imagery.
If NULL, then use the current working directory. The default is NULL.}

\item{nm_mark}{(character) the name mark of clipped images.
The default is "clip". It would be ignored if bry is NULL.}

\item{return_stack}{(logical) if TRUE, stack the imagery together and return.
If the area is large and resolution is high, it is better not to stack them.
The default is TRUE.}
}
\value{
stars if return_stack is TRUE.
The images would be saved as separate files.
}
\description{
This function allows you to parse worldclim version 2.1 files
with a setting of boundary and a few other options.
}
\examples{
worldclim2(var = "tmin", res = 10, return_stack = FALSE)
}
\references{
\url{https://worldclim.org/data/index.html}
}
