% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotree_sdm.R
\name{isotree_po}
\alias{isotree_po}
\title{A function to run extended isolation forest as SDM.}
\usage{
isotree_po(
  occ,
  occ_test = NULL,
  occ_crs = 4326,
  variables,
  ntrees = 100L,
  sample_size = NA,
  sample_rate = NA,
  ndim = 0L,
  ...,
  response = TRUE,
  variable_importance = TRUE,
  simple = FALSE,
  visualize = TRUE
)
}
\arguments{
\item{occ}{(data.frame) The occurrence data frame for training.
There must be column x and y for coordinates. This argument is required}

\item{occ_test}{(data.frame) The occurrence data frame for testing.
There must be column x and y for coordinates if not NULL.
If NULL, no test will be used. The default is NULL.}

\item{variables}{(RasterStack) the stack of environmental variables.
This argument is required.}

\item{ntrees}{(integer) the number of trees for the forest.
The default is 100.}

\item{sample_size}{(integer) a number for sampling size.
If a sample_rate is set, it will be ignored.
The default is NULL.}

\item{sample_rate}{(numeric) a rate for sampling size. Should be in \link{0, 1}.
The default is NULL.}

\item{response}{(logical) if TRUE, generate response curves. The default is TRUE.}

\item{simple}{(logical) if TRUE, make the evaluation a simple data.frame;
if FALSE, make the evaluation a full list. The default is FALSE.}

\item{eval_method}{(character) Method for evaluation.
Should be one of "po" (present-only) and "pa" (present-absence).
The default is po.}

\item{ExtensionLevel}{(integer) ExtensionLevel for isolation forest,
it must be smaller than the dimension of environmental variables.
The default is 0.}

\item{seed}{(integer) set the seed for random process.
The default is 100.}

\item{plots}{(logical) if TRUE, plot out the evaluation figures. The default is TRUE.}

\item{verbose}{(logical) if TRUE, print out the progress. The default is TRUE.}
}
\value{
(eif_sdm) a list of model, occ var_train, pred_train, pred_test,
prediction, responses, eval_train, and eval_test.
}
\description{
This function allows you to call extended isolation forest to do species distribution modeling.
}
\examples{
isotree_po(occ = occ, variables = env_vars, ntrees = 500,
sample_size = 400, ndim = 0)

}
