% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotree_po.R
\name{isotree_po}
\alias{isotree_po}
\title{Function to run extended isolation forest as SDM.}
\usage{
isotree_po(
  occ,
  occ_test = NULL,
  occ_crs = 4326,
  variables,
  categ_vars = NULL,
  ntrees = 100L,
  sample_size = NA,
  sample_rate = NA,
  ndim = 1L,
  seed = 10L,
  ...,
  response = TRUE,
  check_variable = TRUE,
  visualize = FALSE
)
}
\arguments{
\item{occ}{(\code{data.frame}, \code{sf}, \code{SpatialPointsDataFrame})
The occurrence dataset for training.
There must be column \code{x} and \code{y} for coordinates if it is a regular \code{data.frame}.}

\item{occ_test}{(\code{data.frame}, \code{sf}, \code{SpatialPointsDataFrame}, or \code{NULL})
The occurrence dataset for independent test. The same structure as \code{occ}.
If not \code{NULL}, there must be column \code{x} and \code{y} for coordinates when it is a
regular \code{data.frame}. If \code{NULL}, no independent test will be used.
The default is \code{NULL}.}

\item{occ_crs}{(\code{numeric} or \code{\link{crs}}) The EPSG number or
\code{\link{crs}} object of occurrence CRS.
The default value is \code{4326}, which is the geographic coordinate system.}

\item{variables}{(\code{RasterStack} or \code{stars}) The stack of environmental variables.}

\item{categ_vars}{(\code{vector} of \code{character} or \code{NULL}) The names of categorical
variables. Must be the same as the names in \code{variables}.}

\item{ntrees}{(\code{integer}) The number of trees for the isolation forest. It must
be integer, which you could use function \code{\link{as.integer}} to convert to.
The default is \code{100L}.}

\item{sample_size}{(\code{integer} or \code{NULL}) Alternative argument for \code{sample_rate}.
If not \code{NULL}, it should be a number for sampling size in \verb{[2, nrow(occ)]}. It must
be integer, which you could use function \code{\link{as.integer}} to convert to.
The default is \code{NULL}. Only set either \code{sample_size} or \code{sample_rate}.}

\item{sample_rate}{(\code{numeric} or \code{NULL}) Alternative argument for \code{sample_size}.
If not \code{NULL}, it should be a rate for sampling size in \verb{[0, 1]}.
The default is \code{NULL}. Only set either \code{sample_size} or \code{sample_rate}.}

\item{ndim}{(\code{integer}) ExtensionLevel for isolation forest. It must
be integer, which you could use function \code{\link{as.integer}} to convert to.
Also, it must be no smaller than the dimension of environmental variables.
When it is 1, the model is a traditional isolation forest, otherwise the model
is an extended isolation forest. The default is 1.}

\item{seed}{(\code{integer}) The random seed used in the modeling. It should be an
integer. The default is \code{10L}.}

\item{...}{Other arguments that \code{\link{isolation.forest}} needs.}

\item{response}{(\code{logical}) If \code{TRUE}, generate response curves.
The default is \code{TRUE}.}

\item{check_variable}{(\code{logical}) If \code{TRUE}, check the variable importance.
The default is \code{TRUE}.}

\item{visualize}{(\code{logical}) If \code{TRUE}, generate the essential figures related to
the model. The default is \code{FALSE}.}
}
\value{
(\code{POIsotree}) A list of
\itemize{
\item{model (\code{\link{isolation.forest}}) The threshold set in function inputs}
\item{variables (\code{stars}) The formatted image stack of environmental variables}
\item{pts_occ (\code{\link{sf}}) A \code{\link{sf}} of training occurrence dataset}
\item{pts_occ_test (\code{\link{sf}} or \code{NULL}) A \code{\link{sf}} of test occurrence dataset}
\item{var_train (\code{\link{sf}}) A \code{\link{sf}} with values of each
environmental variables for training occurrence}
\item{pred_train (\code{\link{sf}}) A \code{\link{sf}} with values of
prediction for training occurrence}
\item{eval_train (\code{POEvaluation}) A list of presence-only evaluation metrics
based on training dataset. See details of \code{POEvaluation} in \code{\link{evaluate_po}}}
\item{var_test (\code{\link{sf}}) A \code{\link{sf}} with values of each
environmental variables for test occurrence}
\item{pred_test (\code{\link{sf}}) A \code{\link{sf}} with values of
prediction for test occurrence}
\item{eval_test (\code{POEvaluation}) A list of presence-only evaluation metrics
based on test dataset. See details of \code{POEvaluation} in \code{\link{evaluate_po}}}
\item{prediction (\code{stars}) The predicted environmental suitability}
\item{marginal_responses (\code{MarginalResponse}) A list of marginal response values
of each environmental variables. See details in \code{\link{marginal_response}}}
\item{independent_responses (\code{IndependentResponse}) A list of independent response values
of each environmental variables. See details in \code{\link{independent_response}}}
\item{variable_dependence (\code{VariableDependence}) A list of variable dependence values
of each environmental variables. See details in \code{\link{variable_dependence}}}
\item{variable_analysis (\code{VariableAnalysis}) A list of variable importance analysis based on
multiple metrics. See details in \code{\link{variable_analysis}}}}
}
\description{
Call isolation forest and its variations to do species distribution modeling.
}
\details{
Please read details of algorithm \code{\link{isolation.forest}} on
\url{https://github.com/david-cortes/isotree}, and
the R documentation of function \code{\link{isolation.forest}}.
}
\examples{
# Using a pseudo presence-only occurrence dataset of
# virtual species provided in this package
library(dplyr)
library(sf)
library(stars)
library(itsdm)

data("occ_virtual_species")
occ_virtual_species <- occ_virtual_species \%>\%
  mutate(id = row_number())

set.seed(11)
occ <- occ_virtual_species \%>\% sample_frac(0.7)
occ_test <- occ_virtual_species \%>\% filter(! id \%in\% occ$id)
occ <- occ \%>\% select(-id)
occ_test <- occ_test \%>\% select(-id)

env_vars <- system.file(
  'extdata/bioclim_africa_10min.tif',
  package = 'itsdm') \%>\% read_stars() \%>\%
  slice('band', c(1, 5, 12, 16))

mod_virtual_species <- isotree_po(occ = occ, occ_test = occ_test,
  variables = env_vars, ntrees = 200, sample_rate = 0.8, ndim = 2L,
  seed = 123L)

}
\references{
\itemize{
\item{\href{https://cs.nju.edu.cn/zhouzh/zhouzh.files/publication/icd
m08b.pdf?q=isolation-forest}{Liu, Fei Tony, Kai Ming Ting, and Zhi-Hua Zhou.
"Isolation forest."\emph{2008 eighth ieee international conference on data mining}.
IEEE, 2008.}}
\item{\href{https://doi.org/10.1145/2133360.2133363}{Liu, Fei Tony, Kai Ming
Ting, and Zhi-Hua Zhou. "Isolation-based anomaly detection."
\emph{ACM Transactions on Knowledge Discovery from Data (TKDD)} 6.1 (2012): 1-39.}}
\item{\href{https://link.springer.com/content/pdf/10.1007/978-3-642-15883-4_
18.pdf}{Liu, Fei Tony, Kai Ming Ting, and Zhi-Hua Zhou. "On detecting
clustered anomalies using SCiForest." \emph{Joint European Conference on
Machine Learning and Knowledge Discovery in Databases}. Springer, Berlin,
Heidelberg, 2010.}}
\item{\href{https://ieeexplore.ieee.org/stamp/stamp.jsp?arnumber=8888179}{Ha
riri, Sahand, Matias Carrasco Kind, and Robert J. Brunner. "Extended
isolation forest." \emph{IEEE Transactions on Knowledge and Data Engineering (2019)}.}}
\item{\href{https://github.com/david-cortes/isotree}{https://github.com/davi
d-cortes/isotree}}
\item{References of related feature such as response curves and variable importance
will be listed under their own functions}
}
}
\seealso{
\code{\link{evaluate_po}}, \code{\link{marginal_response}}, \code{\link{independent_response}},
\code{\link{variable_dependence}}, \code{\link{variable_analysis}}, \code{\link{isolation.forest}}
}
