% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.POEvaluation}
\alias{plot.POEvaluation}
\title{Show model evaluation.}
\usage{
\method{plot}{POEvaluation}(x, ...)
}
\arguments{
\item{x}{(\code{POEvaluation}) The presence-only evaluation object to plot.
It could be the return of function \code{\link{evaluate_po}}.}

\item{...}{Not used.}
}
\value{
A \code{patchwork} of \code{ggplot2} figure of AUC_ratio, AUC_background and CBI.
}
\description{
Display informative and detailed figures of continuous Boyce
index, AUC curves, and TSS curve.
}
\examples{
\donttest{
# Using a pseudo presence-only occurrence dataset of
# virtual species provided in this package
library(dplyr)
library(sf)
library(stars)
library(itsdm)

# Prepare data
data("occ_virtual_species")
obs_df <- occ_virtual_species \%>\% filter(usage == "train")
eval_df <- occ_virtual_species \%>\% filter(usage == "eval")
x_col <- "x"
y_col <- "y"
obs_col <- "observation"

# Format the observations
obs_train_eval <- format_observation(
  obs_df = obs_df, eval_df = eval_df,
  x_col = x_col, y_col = y_col, obs_col = obs_col,
  obs_type = "presence_only")

env_vars <- system.file(
  'extdata/bioclim_tanzania_10min.tif',
  package = 'itsdm') \%>\% read_stars() \%>\%
  slice('band', c(1, 5, 12, 16))

# With imperfect_presence mode,
mod <- isotree_po(
  obs_mode = "imperfect_presence",
  obs = obs_train_eval$obs,
  obs_ind_eval = obs_train_eval$eval,
  variables = env_vars, ntrees = 20,
  sample_size = 0.8, ndim = 2L,
  seed = 123L, response = FALSE,
  spatial_response = FALSE,
  check_variable = FALSE)

eval_train <- evaluate_po(
  mod$model,
  occ_pred = mod$pred_train$prediction,
  var_pred = na.omit(as.vector(mod$prediction[[1]])))

plot(eval_train)
}

}
\seealso{
\code{\link{evaluate_po}}, \code{\link{print.POEvaluation}}
}
