% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ReducedImageStack}
\alias{print.ReducedImageStack}
\title{Print summary information from ReducedImageStack object.}
\usage{
\method{print}{ReducedImageStack}(x, ...)
}
\arguments{
\item{x}{(ReducedImageStack) A ReducedImageStack object to be messaged.
It could be the return of function \code{dim_reduce}.}

\item{...}{Not used.}
}
\value{
The same object that was passed as input.
}
\description{
Display the most general and informative characteristics of
a ReducedImageStack object.
}
\examples{
print(img_reduced)

}
