% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_analysis.R
\name{variable_analysis}
\alias{variable_analysis}
\title{A function to evaluate relative importance of each variable.}
\usage{
variable_analysis(
  model,
  var_occ,
  var_occ_test = NULL,
  variables,
  shap_nsim = 100,
  visualize = FALSE,
  seed = 10
)
}
\arguments{
\item{model}{(isolation_forest) The extended isolation forest SDM.}

\item{var_occ}{(data.frame, tibble) the data.frame style table that
include values of environmental variables at occurrence locations.}

\item{var_occ_test}{(data.frame, tibble) The data.frame style table that
include values of environmental variables at occurrence locations of testing.
If NULL, no test will be used. The default is NULL.}

\item{visualize}{(logical) if TRUE, plot the response curves.
The default is FALSE.}

\item{permution_cv}{(integer) The time to run permutation method.
The default is 5.}
}
\value{
(VariableAnalysis) a list of variable importance analysis according
to different metrics including jackknife of Pearson correlation and AUC_ratio,
and SHAP test.
}
\description{
This function allows you to evaluate relative importance of
each variable within the model using methods: with this variable only,
with all variables except this variables, and permutation feature importance
method. Because the model is presence-only, we only use pearson correlation
with the result of full model as the metrics to rank the variables.
}
\examples{
variable_analysis(model = isotree_po,
var_occ = var_occ, variables = env_vars)
}
